Config volumes for {oauth2_,hmac,auth}proxy
This makes it possible to update the configuration files on the host machine
without rebuilding the images. It also allows for passing configuration
secrets via files that are mounted from the config directory but listed in
.gitignore so they're not accidentally checked in.

Also adds some new `./go` commands to make it easier to run and experiment
with containers.