Update all configs to use Docker network hostnames
The Docker network 18f/knowledge-sharing-toolkit created by the commands in
the previous commit binds the container names to Docker network addresses via
a local DNS mechanism:

https://docs.docker.com/engine/userguide/networking/dockernetworks/#docker-embedded-dns-server

This means that all of the configurations need to be updated to reference
specific container "hosts", rather than "localhost" or "127.0.0.1". The nice
feature is that the dependencies between services become more explicit via the
configuration files.